requirements.txt
streamlit
pandas
pandas_datareader
plotly
numpy
Pillow
scikit-learn
scipy
sklearn
yfinance
mplfinance
altair